Desktop. Menu. Support shortcuts and mnemonics for menu items
KeyShortcut is implemented as a top-level data class, because it not only represent
shortcut for menu, but shortcut in general.

In the future we can reuse this data class for other purposes
(define local shortcuts, for example).

Currently Swing can't support shortcuts with arbitrary keys,
it supports only shortcut with single key and multiple modifiers.
It isn't possible, because pointer event doesn't contain all currently pressed keys,
it only contain currently pressed modifiers.

When we will support knowing the current pressed keys,
we can support arbitrary keys in KeyShortcut (simply extending the API).

Change-Id: Icabf562ac72514bcd6a29649d0fa856689c5e4ec
Test: manual ./gradlew :compose:desktop:desktop:desktop-samples:runWindowApi -Pandroidx.compose.multiplatformEnabled=true (menu example)