Fix dismissAllowingStateLoss() when on the back stack
When a DialogFragment is shown via the
show(FragmentTransaction, String) method, it is
added to the FragmentManager's back stack. As
FragmentManager does not allow calling
popBackStack() while allowing state loss, this
causes issues when using the DialogFragment
API of dismissAllowingStateLoss().

As using the 'AllowingStateLoss()' APIs are always
a bad idea (checking for isStateSaved() is the
better choice in all cases, but not running code
after your state is saved is even better), we
don't want to add any new public API to support
this use case. Instead, we'll pass through the
allowStateLoss value through to FragmentManager
for just this particular case.

Relnote: "The `dismissAllowingStateLoss()` method
of `DialogFragment` no longer crashes when you call
it after the state is saved when specifically adding
the DialogFragment via the
`show(FragmentTransaction, String)` method."
BUG: 27374448
Test: new DialogFragmentTest tests pass

Change-Id: I844222aba9e56d7ae6b0cdb57793ab38f6dbe4c1