Updates SampledAnnotationEnforcer to work with partial analysis
Lint from CLI now only runs in partial analysis mode, so we now build a graph of @sample links and @Sampled functions, and merge them together during reporting.

Also removes some @Sampled annotations from the datastore-sampleapp that are not being referenced in documentation, so are not needed (and would cause lint errors if not removed).

Fixes: b/181900386
Fixes: b/187418637
Test: lintDebug
Change-Id: I2978b3bf8320860087b7c34b9b3dc8b5908e5080