AndroidX Webkit: remove extra quote char
No change to logic, only a documentation change. This deletes an extra
double-quote character we found in the API doc for WebView dark theme
strategies. This extra quote character was a typo and should be removed.

I've double-checked the example "<meta>" tag looks correct.

Bug: 194343633
Test: None
Relnote: Fixed a typo in dark theme documentation.
Change-Id: I36ebf63c24427e8a19a3ef3d202a5a277d52cc16