Fix invalidation checks for DataSource.isInvalid
aosp/1763933 added a behavior change to immediately trigger invalidate
callbacks if the source they're registered on is already invalid to fix
cases where a PagedListBuilder could get stuck due to a race between the
invalidation check, when the datasource was actually invalidated and
when the invalidation callback was registered.

This CL addresses a bug introduced by aosp/1763933 that does not account
for overrides of DataSource.isInvalid and removes a redundant check in
LegacyPagingSource which would call .invalidate() an extra time.

Test: ./gradlew paging:paging-common:test
Change-Id: Icd8f7f7c4f63062f1e6f606b8ce38687d184f22b