Desktop. Key handlers in dialog didn't work
The bug was because AWT blocks the main thread when we call `isVisible = true`
(and creates an internal event loop)

So our code after `isVisible = true` never executed.
Moved this code into addNotify - the method which will be called when the dialog will be initializated.

Also copy-pasted WindowTest to DialogTest.
I thought about extracting the base Test class, but realized that it is better to keep them stupid, even with copy-paste.
Because the tests looks clener that way, without additional abstractions

Test: ./gradlew jvmTest desktopTest -Pandroidx.compose.multiplatformEnabled=true
Change-Id: I522c19419a901dfaa6317a72af038a8c3236ef4f