Workaround KSP enum annotation value inconsistency.
When reading an annotation value that is an enum entry (or array of enum entries), KSP will return different objects in its KSValueArgument API depending if the annotation being read is from compiled code or not. Specifically it returns a KSType when its from source and the actual JavaEnum when compiled code. See https://github.com/google/ksp/issues/429

This change workarounds the issue by wrapping the value in a KspEnumEntry for XAnnotationValue and by not using valueOf() for XAnnotationBox.

Test: XAnnotationTest.kt and XAnnotationBoxTest.kt
Change-Id: Ie64c08dbf2a5d090b0c7b347b24558d039ab39c5