Ensure we keep track of all transitioning entries
We currently keep up with all transitioning entries in a set, which
means we can't keep up with multiple entries that are considered "equal"
but actually are different objects.

This currently breaks navigate with popUpTo and transitions since we
will pop first, causing us to consider the old entry as transitioning,
then when we navigate we attempt to add the new entry as transitioning,
but since it is "equal" to the old one, the set just ignores the add.

We need to instead make transitionsInProgress a List, meaning we can
have multiple entries that "equal" each other. This also means we need
to be sure to clear all the "equal" entries from the list whenever one
of them is complete.

Relnote: "Re-selecting the same tab when using navigation compose with
multiple back stacks will no longer result in a blank screen."

Test: added testSameEntry
Bug: 194925622

Change-Id: I860dc9f2cc2516924c03dba491a37aa8ace99bb3