Fix bug causing multiple underlines to be applied in certain situations.
In certain situations (e.g. when mixing two different scripts in the
same Text element), two underlines could be renderered instead of one if
the element was set to underline. This has been fixed by using
UnderlineSpan instead of setting the "underline" bit on the TextPaint.

Bug: 195388378
Test: New screenshot tests added.
Relnote: "Fix bug causing multiple underlines to be applied in certain
situations."

Change-Id: Ib6712e68987be272f471e8c229840f57cd6f3845