Update KSP to beta07
Moved to the new getThrownTypes API for accessors instead of the custom
implementation. Unfortunately, it is broken for pre-compiled code.

https://github.com/google/ksp/issues/505

Similarly, I've tried enabling the XExecutableElementTest.thrownTypes
test for thrown types in compiled code but unfortunately, it still fails
for thrown types in constructors.
https://github.com/google/ksp/issues/507

I've not cleaned other already fixed issues yet since we've been falling
behind in KSP versions and updating testing is important for Dagger to
use XProcessing.

Also updated the paging compose sample and the leanback sample which were
depending on a predefined version of room instead of projectOrArtifact.
Using predefined version would require keeping ksp prebuilts so i decided
to make it use tip of tree when available. Unfortunately, this required
adding room to the compose build but seems like the right thing to do
for now.

Bug: n/a
Test: ./gradlew :room:test
Change-Id: I0a4f2c0f6f83866eb2c412a0eb8d306789ccc53f