Fix possible flickering on CurvedRow
Turns out Compose may call the placementBlock passed in the layout
without doing a recomposition (i.e., calling the outer function first),
so we can't initialize a variable in the main code of the function that
is then updated in the placementBlock.
Also, ensure that we remeasure the CurvedRow when the intrinsic
dimensions of the BasicCurvedText change (for example, because we
changed the text).

Bug: 196829063
Test: Added
Relnote: "Fix possible flickering on CurvedRow when the content updates,
and ensures the curved row is remeasured"

Change-Id: Ie4e06fa0f9a7cc74b5c5a22cdc133091a6591990