Add Background modifier, along with the Color classes.
This is currently scoped to Wear as we will extend this with background
shape in future, which is not easily done for AppWidgets on API < 31.
While we figure out how to support that, we'll keep it as part of Wear,
then move across in future if it makes sense.

Relnote: n/a
Test: Included.
Change-Id: I3dbaba21272876d42aa009a66c4327c2b0a5e3e4