Use support root folder for project path validation
This CL makes project directory validation compatible with playground
projects.

Previously, it was comparing against the root project directory, which
is different for playground projects. I think it means to check against
the support root folder, hence this CL updates the check to use support
root directory (instead of root project directory) to make it compatible
with playground projects.

Bug: 191911288
Test: cd room; ./gradlew studio runs properly.

Change-Id: I6c3cf01826b7218c00acafa0fb9291ba244ab656