TimeText implementation for wear compose
TimeText composable helps developers to show a clock on square and round devices.
If screen is square then the clock text is shown horizontally. If screen is round - text goes along the edge of the screen.
TimeText supports a lot of customization like leading and trailing texts, separators, colors, etc. Basically every additional content is a slot, so can be easily customized.
The only problem with this implementation is that CurvedRow can't "bend" composables, eg we can't use Text on round devices, because composable itself can't be bended ( or it's too expensive and ugly to consider this option). For that reason we use 2 types of slots - one for square devices with Text ( or other composable) and another for round devices ( with CurvedRowScope).
That's the tradeoff which we had to make to leave this api reasonably understandable and at the same time customizable.

Bug: 196200654
Relnote: "Added TimeText implementation for Wear Compose"
Test: Run tests in androidx.wear.compose.integration.demos.test
Change-Id: I5654cc5c2e02a3eff5fd48317b5a8600d3b17644