Do not track remote states locally within PageFetcher for PageEvents
Previously from fetcher side, we locally tracked remote load states
per-generation to inject them into the page event stream, however this
leads to issues between generations, where the local state is incorrect
before it receives an event from either the source or remote stream.

This CL changes .injectRemoteEvents into a custom .combine operator so
that we always wait for a real event from both remote and source flows,
so that we are guaranteed to have valid states for both. We also now
immediately send REFRESH=Loading from source so that it is less likely
to drop remote events due to waiting on source flow from .combine.

The reason we use a custom .combine operator is to remove the "batching"
logic in .combine, which would normally output the latest values from
each Flow and can miss emissiosn when we get multiple emissions on the
same Flow, as opposed to alternating emissions.

Since it also affected some tests, this CL also fixes a longstanding
issue with cancellation on invalidation. Previously, a collector
would continue to receive remote events after invalidation, which
can make it difficult for Java users or those not using collectLatest to
switch to the next generation, since submitData would never return until
cancelled. We now eagerly cancel collection on the remote stream (within
a generation) to close the Flow<PageEvent> from a PagingData as soon as
we receive a new generation.

RelNote: "LoadStates from Paging now await valid values from both
  PagingSource and RemoteMediator before emitting downstream between
  generations. This prevents new generations of PagingData from sending
  NotLoading in CombinedLoadStates.source.refresh if it was already
  Loading; new generations of PagingData will now always correctly begin
  with Loading for refresh state instead of first resetting to NotLoading
  incorrectly in some cases.

  Cancellation on past generations now happen eagerly on invalidation /
  new generations. It should no longer be required to use .collectLatest
  on Flow<PagingData>, although it is still highly recommended to do so."

Fixes: 177351336
Fixes: 195028524
Test: ./gradlew paging:paging-common:test
Change-Id: I0b2b5e0120e2e3a677f65d40d3955a9bbbe2e1a9