Ensure that add+remove cleans up the Fragment state
When using setReorderingAllowed(true), it is possible
for a Fragment to be both added to the FragmentManager
and marked as removed before it moves through any
lifecycle changes. This means that the work normally
done in destroy() and detach() doesn't happen so we
need to manually do that work if the state hasn't changed.

Test: newly added test passes
Relnote: "Fixed an issue where the state of a Fragment
that was added with `setReorderingAllowed(true)` and then
immediately removed before executing pending transactions
would not be properly cleaned up."

Change-Id: I8ccb834a4337dda503e64402dab4aca807d81788