Add function for arguments and deeplinks in nested graphs
Navigation compose's composable destination offers the ability to take a
list of arguments and/or deep links and add them all to your
NavDestination. For nested navigation graphs, you still have to use the
builder functions.

It would be nice if we can add a way for nested graphs to also take a
list so that the APIs are uniform.

RelNote: "You can now pass in a list of arguments and/or deep links to your
nested navigation graph's builder and they will automatically be added
to the resulting graph."
Test: added tests
Bug: 182545357

Change-Id: I8a47074b68f36d9195f353ef7c726c75b8e1a73f