Add <queries> tag for launcher items.
The watch faces need to access the launcher items to support the
standard app launcher complication. This requires, with the new
package visibility restrictions, for this to be declared in the
manifest.

As part of that, add SMALL_IMAGE to the list of support types by the
complication sample so that this can be tested end-to-end.

Bug: 199246800
Test: manual
Change-Id: Ia2f19326a920f051d1402e7023cf4ab15f5abd49