Restore navigator state before the back stack
When we restore DynamicNavGraphs, the need to be inflated before we can
restore the back stack. This is only done as part of onRestoreState()
when we call replaceWithIncludedNav. Failing to do so means that the
NavGraph's startDestination does not exist when restoring the back
stack.

RelNote: "Dynamic Navigation will no longer crash when the Activity is
recreated."
Test: added include test
Bug: 197942869

Change-Id: Ifa6a3e2f563bbe181cb1fdfb8460217e77bb9338