Cleanup: Remove useless setAttachedUseCasesUpdateListener related code
This was used to check whether a necessary Preview or ImageCapture with matching effect mode is bound together to make extensions work well. After lagecy extensions API was removed and aosp/1749401 was merged, an extra Preview or ImageCapture will be bound automatically. The setAttachedUseCasesUpdateListener related code is useless and can be rmeoved.

Bug: 154959427
Test: ./gradlew bOS
Change-Id: Ia7618ec94ffd8ffce07d6156fe2329a5641afe6e