Add validating API to XElement
This allows processor authors to better handle missing types, possibly deferring elements as needed.

In javac, visitors tend to throw TypeNotPresentException, this CL also updates JavacAnnotationValue to do the same instead of creating annotation values with '<error>' elements which can be harder to deal with. The idea is that a processor author should validate their elements as necessary which would avoid a TypeNotPresentException when further inspecting them.

Also removed workaround regarding validating locals.

Test: XProcessingStepTest
Change-Id: Ibcc95608f5439d32a062c7590563a4b5062cceef