Initial impl of StartupTimingQuery, part of new startup metric
Bug: 183129298
Bug: 190219056
Test: ./gradlew benchmark:benchmark-common:cC benchmark:benchmark-macro:cC

Tested only on API 31 so far, as there are issues with capturing
traces (with app trace tags enabled) on older platform versions.

Adds check for atrace args to be too long on older platform versions,
which was causing problems when attempting to test on API 23.

Known compatibility issues for lowering min API (e.g. 'launching'
missing for hot launches on certain old versions) will be handled in
follow up work.

Change-Id: I2eb4700f2c7cbc822ef5a23da7b97192da0ee177