Add jbpresubmit.yml to run the workflow in jb androidx fork
This workflow is intended to be used by JB-compose team in https://github.com/JetBrains/androidx
which is a fork of androidx repo.
By having this file in the AndroidX upstream we (JB-compose team) avoid the need to
rebase/cherry-pick the workflow file for every CL we want to submit upstream.
The aim of this workflow is to run checks and some tests before submitting the CL upstream.

Test: we run the workflow in jb androidx fork

Change-Id: If0442c317d26fb432b19845cd0b8873da20bc794