Add a Picker component to select from a list of options.
Adds a picker component that lets the user select a item from a
scrolling list. By default, the list of selectable items is repeated
'infinitely' in both directions, to give the impression of a rotating
cylinder seen from the side.
There is two key things that will be added later since they require
deeper changes in ScalingLazyColumn and/or LazyColumn:
* Snapping to a value after a swipe/fling.
* Adding a function to the state to set/scroll to the current value.

Bug: 199476864
Test: Some added, more to come when we have better control.
Relnote: "Add a Picker component to select from a list of options."

Change-Id: I6461b07145e55d7e3d5f9db6653227d0f9670d75