Include new api to hide a shortcut from launcher
Include a new field in ShortcutInfoCompat which serves as an indication
of whether a shortcut is hidden from launcher. Shortcut marked as hidden
from launcher will not be included in the search result in
LauncherApps#getShortcuts. This generally means the shortcut would not
be displayed by a launcher app (e.g. Long-Press menu), while remain
visible in other surfaces such as assistant or on-device-intelligence.

Relnote: "Added a new API
`ShortcutInfoCompat.Builder#setHiddenFromLauncher`
to determine whether the shortcut should be hidden from launcher."
Bug: 202335257
Test: ShortcutManagerCompatTest

Change-Id: Ia2a7147fab366308c392eb3dcf0cc6122d7f6181