Fix ConfirmationOverlay when it has no message.
This previously set the TextView to be GONE if it had no message. This
is problematic after aosp/1818161, which used a minHeight constraint on
the TextView to push the image into the correct location (and thus,
without any text, causing the image to float to the bottom of the
screen).

This patch removes this code; the TextView is now always visible
(although we could equally mark it as INVISIBLE if we wanted), so it
pushes the image into the correct location.

Test: Manually verified.
Bug: 203534227
Change-Id: I63e6f0d8b6a310a9f2f642d4fdc7ec4b09f147af